+++ .gitignore
@@ -3,3 +3,5 @@ cyape-1.0
 *.mp3
 latex
 inc
+unit
+scratchpad
