--- src/cyape.h
@@ -1,9 +1,0 @@
-#ifndef __CYAPE_H__
-#define __CYAPE_H__
-//Generated with genfile -d cyape.h
-#include "include/common.h"
-#include "include/graphics.h"
-#include "include/audio.h"
-#include "include/net.h"
-
-#endif //__CYAPE_H__
