+++ src/include/gobjects/ai/ai.h
@@ -1,0 +1,10 @@
+#ifndef AI_H
+#define AI_H
+#include "../../common.h"
+#include "../sprite.h"
+#include "../../graphics.h"
+//Generated with genfile -d ai.h -i "['../../common.h', '../sprite.h', '../../graphics.h']"
+
+
+
+#endif //AI_H
