+++ src/include/gobjects/sprite.h
@@ -1,0 +1,9 @@
+#ifndef SPRITE_H
+#define SPRITE_H
+#include "../common.h"
+#include "../graphics.h"
+//Generated with genfile -d sprite.h -i "['../common.h', '../graphics.h']"
+
+
+
+#endif //SPRITE_H
