+++ src/include/graphics/label.h
@@ -1,0 +1,15 @@
+#ifndef LABEL_H
+#define LABEL_H
+#include "../common.h"
+#include "texture.h"
+//Generated with genfile -d label.h -i "['../common.h']"
+
+
+struct y_label {
+
+};
+
+typedef struct y_label y_label;
+
+
+#endif //LABEL_H
