+++ src/include/graphics/menu.h
@@ -1,0 +1,8 @@
+#ifndef MENU_H
+#define MENU_H
+#include "../common.h"
+//Generated with genfile -d menu.h -i "['../common.h']"
+
+
+
+#endif //MENU_H
