+++ src/include/graphics/texture.h
@@ -1,0 +1,14 @@
+#ifndef TEXTURE_H
+#define TEXTURE_H
+#include "../common.h"
+//Generated with genfile -d texture.h -i "['../common.h']"
+
+
+struct y_img{
+
+};
+
+typedef struct y_img y_img;
+
+
+#endif //TEXTURE_H
