--- src/include/input.c
@@ -1,3 +1,0 @@
-#include "input.h"
-//Generated with genfile -c input.c -i "input.h"
-
