--- src/include/input.h
@@ -1,8 +1,0 @@
-#ifndef __INPUT_H__
-#define __INPUT_H__
-#include "common.h"
-//Generated with genfile -d input.h -i "['common.h']"
-
-
-
-#endif //__INPUT_H__
