--- src/include/net.c
@@ -1,4 +1,0 @@
-#include "net.h"
-//Generated with genfile -c net.c -i "net.h"
-
-
