--- src/include/physics.c
@@ -1,9 +1,0 @@
-#include "physics.h"
-//Generated with genfile -c physics.c -i "physics.h"
-
-
-int main(int argc, char * argv[])
-{
-
-    return 0;
-}
