+++ src/include/yape.h
@@ -1,0 +1,11 @@
+#ifndef YAPE_H
+#define YAPE_H
+#include "common.h"
+#include "graphics.h"
+#include "audio.h"
+#include "physics.h"
+//Generated with genfile -d yape.h -i "['graphics.h', 'audio.h', 'physics.h']"
+
+
+
+#endif //YAPE_H
