--- src/main.h
@@ -1,3 +1,0 @@
-#include "cyape.h"
-
-void hollaback_yall(y_texture * hresult, SDL_Renderer * hrnd);
