+++ src/unit/main.c
@@ -1,0 +1,12 @@
+#include "main.h"
+
+
+
+int main(int argc, char ** argv){
+	y_pair test = { 0,0 };
+	y_pair best = { 100,100 };
+
+	printf("test: %d, %d\nbest: %d,%d\n",test.x,test.y,best.x,best.y);
+
+	return 0;
+}
