+++ src/unit/main.h
@@ -1,0 +1,8 @@
+#ifndef MAIN_H
+#define MAIN_H
+#include "../include/common.h"
+//Generated with genfile -d main.h -i "['../include/common.h']"
+
+
+
+#endif //MAIN_H
